pyproject.toml
[tool.poetry]
name = "code-wizard"
version = "0.1.0"
description = ""
authors = ["RutamBhagat <rutambhagat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
openai = "^1.23.2"
beautifulsoup4 = "^4.12.3"
black = "^24.4.0"
tiktoken = "^0.6.0"
pinecone-client = "^3.2.2"
unstructured = "^0.13.3"
nltk = "^3.8.1"
fastapi = "^0.110.2"
streamlit = "^1.33.0"
streamlit-chat = "^0.1.1"
tqdm = "^4.66.2"
langchain-pinecone = "^0.1.0"
python-dotenv = "^1.0.1"
nbconvert = "^7.16.3"
unstructured-client = "^0.22.0"
markdown = "^3.6"
transformers = "^4.40.1"
uvicorn = "^0.29.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
